Casting properly

+++ src/darwin/Framework/CHIP/MTRDevice.mm
@@ -293,11 +293,11 @@ typedef NS_ENUM(NSUInteger, MTRDeviceWorkItemDuplicateTypeID) {
     _state = state;
     if (lastState != state) {
         if (state != MTRDeviceStateReachable) {
-            MTR_LOG_INFO("%@ State change %lu => %lu, set estimated start time to nil", self, lastState, state);
+            MTR_LOG_INFO("%@ State change %lu => %lu, set estimated start time to nil", self, (unsigned long)lastState, (unsigned long)state);
             _estimatedStartTime = nil;
             _estimatedStartTimeFromGeneralDiagnosticsUpTime = nil;
         } else {
-            MTR_LOG_INFO("%@ State change %lu => %lu", self, lastState, state);
+            MTR_LOG_INFO("%@ State change %lu => %lu", self, (unsigned long)lastState, (unsigned long)state);
         }
         id<MTRDeviceDelegate> delegate = _weakDelegate.strongObject;
         if (delegate) {
